+++ src/prefect/core/task.py
@@ -129,19 +129,19 @@ class Task(metaclass=SignatureValidator):
                 regardless of trigger. By default, this prevents tasks from attempting to use either state or data
                 from tasks that didn't run. If `False`, the task's trigger will be called as normal,
                 with skips considered successes. Defaults to `True`.
-        - cache_for (timedelta, optional): The amount of time to maintain a cache
+        - cache_for (timedelta, optional, DEPRECATED): The amount of time to maintain a cache
             of the outputs of this task.  Useful for situations where the containing Flow
             will be rerun multiple times, but this task doesn't need to be.
-        - cache_validator (Callable, optional): Validator that will determine
+        - cache_validator (Callable, optional, DEPRECATED): Validator that will determine
             whether the cache for this task is still valid (only required if `cache_for`
             is provided; defaults to `prefect.engine.cache_validators.duration_only`)
-        - cache_key (str, optional): if provided, a `cache_key` serves as a unique identifier for this Task's cache, and can
+        - cache_key (str, optional, DEPRECATED): if provided, a `cache_key` serves as a unique identifier for this Task's cache, and can
             be shared across both Tasks _and_ Flows; if not provided, the Task's _name_ will be used if running locally, or the
             Task's database ID if running in Cloud
         - checkpoint (bool, optional): if this Task is successful, whether to
             store its result using the `result_handler` available during the run; Also note that
             checkpointing will only occur locally if `prefect.config.flows.checkpointing` is set to `True`
-        - result_handler (ResultHandler, optional): the handler to use for
+        - result_handler (ResultHandler, optional, DEPRECATED): the handler to use for
             retrieving and storing state results during execution; if not provided, will default to the
             one attached to the Flow
         - state_handlers (Iterable[Callable], optional): A list of state change handlers
@@ -247,6 +247,18 @@ class Task(metaclass=SignatureValidator):
         self.cache_validator = cache_validator or default_validator
         self.checkpoint = checkpoint
         self.result_handler = result_handler
+
+        if cache_for or cache_key or cache_validator:
+            warnings.warn(
+                "DEPRECATED: all cache_* Task options are deprecated, these functionalities have been moved to the prefect.engine.Result object.",
+                UserWarning,
+            )
+
+        if result_handler:
+            warnings.warn(
+                "DEPRECATED: the result_handler Task option is deprecated, this is being replaced by the prefect.engine.Result object.",
+                UserWarning,
+            )
 
         if state_handlers and not isinstance(state_handlers, collections.Sequence):
             raise TypeError("state_handlers should be iterable.")